Finish facet tree creation, hook into rake import task.

+++ app/models/ncbi_species_terms.rb
@@ -11,28 +11,42 @@ class NcbiSpeciesTerms < ActiveRecord::Base
 
   def self.load_from_tree_file(tree_filename)
     entries = []
+    count = 0
     File.open(tree_filename) do |f|
       f.each do |line|
         fields = line.strip.split('|')
-        entries << NcbiSpeciesTerms.new(species_taxon_id: fields[0],
-                                        term: fields[1],
-                                        term_type: fields[2],
-                                        full_tree_id: fields[3]
-                                       )
+        entries << fields[0..3]
+        #entries << NcbiSpeciesTerms.new(species_taxon_id: fields[0],
+        #                                term: fields[1],
+        #                                term_type: fields[2],
+        #                                full_tree_id: fields[3]
+        #                               )
+        count += 1
+        if (count % 10000) == 0
+          print "importing #{count - 999}--#{count}."
+          NcbiSpeciesTerms.import [:species_taxon_id, :term, :term_type, :full_tree_id], entries
+          print ".\n"
+          entries = []
+        end
       end
     end
-    NcbiSpeciesTerms.import entries
+    print "importing #{count}."
+    NcbiSpeciesTerms.import [:species_taxon_id, :term, :term_type, :full_tree_id], entries
   end
 
   def self.generate_facet_treenumbers(&block)
-    xform = TreeTransform.new
-    yield xform
-    NcbiSpeciesTerms.each do |term|
+    tt = TreeTransform.new
+    yield tt
+    NcbiSpeciesTerms.find_each do |term|
       facet_tree = nil
-      if ["species"].include? term.rank
-        facet_tree = xform(term.full_tree_id)
+      # only give facets to species terms
+      if ["species", "species group", "species subgroup", "subspecies"].include? term.term_type
+        facet_tree = tt.transform(term.full_tree_id)
       end
-      term.facet_tree = facet_tree
+      if term.facet_tree_id != facet_tree
+        term.facet_tree = facet_tree
+        term.save
+      end
     end
   end
 
@@ -68,7 +82,6 @@ class NcbiSpeciesTerms < ActiveRecord::Base
         unless i.nil?
           elements = elements[i..-1]
           ok = true
-          break
         end
       end
       return nil unless ok
@@ -77,24 +90,28 @@ class NcbiSpeciesTerms < ActiveRecord::Base
       @prune_nodes.each do |taxid|
         return nil if elements.include? taxid
       end
-      
-      # does this contain any nodes to be removed?
+
+      # are there any nodes to be removed?
       @remove_nodes.each do |taxid|
         elements.delete(taxid)
       end
 
-      ranks = get_ranks(elements)
+      # remove any nodes which have forbidden ranks
+      if @remove_ranks.length > 0
+        ranks = get_ranks(elements)
 
-      @remove_ranks.each do |rank|
-        loop do
-          i = ranks.index(rank)
-          break if i.nil?
-          ranks.delete_at(i)
-          elements.delete_at(i)
+        @remove_ranks.each do |rank|
+          loop do
+            i = ranks.index(rank)
+            break if i.nil?
+            ranks.delete_at(i)
+            elements.delete_at(i)
+          end
         end
       end
 
       # done!
+      return nil if elements.length == 0
       elements.join(".")
     end
 
@@ -103,8 +120,8 @@ class NcbiSpeciesTerms < ActiveRecord::Base
       taxid_list.map do |taxid|
         r = @rank_cache[taxid]
         if r.nil?
-          term = NcbiSpeciesTerms.find(taxid)
-          r = term.term_type unless term.nil?
+          term = NcbiSpeciesTerms.find_by_species_taxon_id(taxid)
+          @rank_cache[taxid] = r = term.term_type unless term.nil?
         end
         r
       end
